refactor:  전화번호 인증 앱 분리
