+++ user_demo/phone_verify/admin.py
@@ -1,0 +1,3 @@
+from django.contrib import admin
+
+# Register your models here.
