+++ user_demo/phone_verify/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class PhoneVerifyConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "phone_verify"
