+++ user_demo/user_demo/settings.py
@@ -45,6 +45,7 @@ INSTALLED_APPS = [
     "rest_framework",
     "rest_framework_simplejwt",
     "users",
+    "phone_verify",
 ]
 
 MIDDLEWARE = [
