+++ user_demo/users/serializers.py
@@ -1,7 +1,7 @@
 from django.core.validators import RegexValidator
 from rest_framework import serializers
 
-from users.models import User, UserVerify
+from users.models import User
 
 
 class UserSerializer(serializers.ModelSerializer):
@@ -28,29 +28,6 @@ class UserDetailSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
         fields = ["email", "name", "nickname", "phone_number"]
-
-
-class UserVerifyCreateSerializer(serializers.ModelSerializer):
-    # 전화번호 형식 벨리데이터
-    phoneNumberRegex = RegexValidator(
-        regex=r"^01([0|1|6|7|8|9]?)-?([0-9]{3,4})-?([0-9]{4})$"
-    )
-    phone_number = serializers.CharField(validators=[phoneNumberRegex])
-
-    class Meta:
-        model = UserVerify
-        fields = ["phone_number"]
-
-
-class UserVerifySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = UserVerify
-        fields = ["key", "phone_number", "is_verified", "token"]
-        extra_kwargs = {
-            "key": {"write_only": True},  # 반환값에서 key는 반환하지 않도록 설정
-            "is_verified": {"read_only": True},
-            "token": {"read_only": True},
-        }
 
 
 class UserLoginSerializer(serializers.Serializer):
