+++ user_demo/users/tests.py
@@ -1,6 +1,7 @@
 from django.test import TestCase
 
-from users.models import User, UserVerify
+from users.models import User
+from phone_verify.models import UserVerify
 from rest_framework_simplejwt.tokens import RefreshToken, AccessToken
 
 # Create your tests here.
